refactor: constants 폴더 분리
