+++ src/js/constants/attribute.js
@@ -1,9 +1,3 @@
-export const LOTTO_RULE = {
-  MIN_NUMBER: 1,
-  MAX_NUMBER: 45,
-  NUMBERS_COUNT: 6,
-};
-
 export const ID = {
   APP: 'app',
   PURCHASE_FORM: 'purchase-form',
@@ -32,22 +26,4 @@ export const CLASS = {
   UNFOLD: 'unfold',
   OPEN: 'open',
   SCROLL_BLOCK: 'scroll-block',
-};
-
-export const LOTTO_PRICE = 1000;
-
-export const MAX_PURCHASABLE_CASH = 100000;
-
-export const ALERT_MESSAGE = {
-  NOT_DIVISIBLE: `${LOTTO_PRICE.toLocaleString()}으로 나누어떨어지는 금액을 입력해주세요.`,
-  DUPLICATED_NUMBER: '당첨 번호가 서로 중복되지 않게 입력해주세요.',
-  OVER_MAX_CASH: `최대 ${MAX_PURCHASABLE_CASH.toLocaleString()}원까지 구매 가능합니다.`,
-};
-
-export const WINNING_PRIZE = {
-  '3개': 5000,
-  '4개': 50000,
-  '5개': 1500000,
-  '5개+보너스볼': 30000000,
-  '6개': 2000000000,
 };
