+++ src/js/constants/lotto.js
@@ -1,0 +1,17 @@
+export const LOTTO_RULE = {
+  MIN_NUMBER: 1,
+  MAX_NUMBER: 45,
+  NUMBERS_COUNT: 6,
+};
+
+export const LOTTO_PRICE = 1000;
+
+export const MAX_PURCHASABLE_CASH = 100000;
+
+export const WINNING_PRIZE = {
+  '3개': 5000,
+  '4개': 50000,
+  '5개': 1500000,
+  '5개+보너스볼': 30000000,
+  '6개': 2000000000,
+};
